requirements_clip.txt
torch
clip @ git+https://github.com/openai/CLIP.git
numpy
Pillow
tqdm 